// client/da/uml-server/src/language-server/generated/ast.ts
export type Entity = Class | DataType | Enumeration | InstanceSpecification | Interface | Package | PrimitiveType;

export const Entity = 'Entity';
// ...
export type Visibility = 'PACKAGE' | 'PRIVATE' | 'PROTECTED' | 'PUBLIC';

export function isVisibility(item: unknown): item is Visibility {
    return item === 'PUBLIC' || item === 'PRIVATE' || item === 'PROTECTED' || item === 'PACKAGE';
}

export interface Class extends AstNode {
    readonly $type: 'AbstractClass' | 'Class';
    __id: string
    isAbstract: boolean
    isActive: boolean
    name: string
    operations: Array<Operation>
    properties: Array<Property>
    skip: boolean
    visibility: Visibility
}

export const Class = 'Class';

export function isClass(item: unknown): item is Class {
    return reflection.isInstance(item, Class);
}

export interface ClassDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'ClassDiagram';
    __id: string
    diagramType: 'CLASS'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const ClassDiagram = 'ClassDiagram';

export function isClassDiagram(item: unknown): item is ClassDiagram {
    return reflection.isInstance(item, ClassDiagram);
}

export interface DataType extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'DataType';
    __id: string
    isAbstract: boolean
    name: string
    operations: Array<Operation>
    properties: Array<Property>
    visibility?: Visibility
}

export const DataType = 'DataType';

export function isDataType(item: unknown): item is DataType {
    return reflection.isInstance(item, DataType);
}

export interface Diagram extends AstNode {
    readonly $type: 'Diagram';
    diagram: UnionType_0
    metaInfos: Array<MetaInfo>
}

export const Diagram = 'Diagram';

export function isDiagram(item: unknown): item is Diagram {
    return reflection.isInstance(item, Diagram);
}

export interface Enumeration extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Enumeration';
    __id: string
    name: string
    values: Array<EnumerationLiteral>
}

export const Enumeration = 'Enumeration';
// ...
export interface InstanceSpecification extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'InstanceSpecification';
    __id: string
    name: string
    slots: Array<Slot>
    visibility?: Visibility
}

export const InstanceSpecification = 'InstanceSpecification';

export function isInstanceSpecification(item: unknown): item is InstanceSpecification {
    return reflection.isInstance(item, InstanceSpecification);
}

export interface Interface extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Interface';
    __id: string
    name: string
    operations: Array<Operation>
    properties: Array<Property>
}

export const Interface = 'Interface';
// ...
export interface Package extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'Package';
    __id: string
    entities: Array<Entity>
    name: string
    uri?: string
    visibility?: Visibility
}

export const Package = 'Package';

export function isPackage(item: unknown): item is Package {
    return reflection.isInstance(item, Package);
}

export interface PackageDiagram extends AstNode {
    readonly $container: Diagram;
    readonly $type: 'PackageDiagram';
    __id: string
    diagramType: 'PACKAGE'
    entities: Array<Entity>
    relations: Array<Relation>
}

export const PackageDiagram = 'PackageDiagram';
// ...
export interface PrimitiveType extends AstNode {
    readonly $container: ClassDiagram | Package | PackageDiagram;
    readonly $type: 'PrimitiveType';
    __id: string
    name: string
}

export const PrimitiveType = 'PrimitiveType';
// ...
export interface Relation extends AstNode {
    readonly $type: 'Abstraction' | 'Aggregation' | 'Association' | 'Composition' | 'Dependency' | 'Generalization' | 'InterfaceRealization' | 'PackageImport' | 'PackageMerge' | 'Realization' | 'Relation' | 'Substitution' | 'Usage';
    __id: string
    relationType: RelationType
    source: Reference<Entity>
    target: Reference<Entity>
}

export const Relation = 'Relation';